Add ability to add a comment when assigning the PR
This change converts the project to typescript and also adds the ability
to add a comment to the PR when assigning the configured reviewers.

This is a backwards incompatible change because of the change in
configuration format.
